Hformer/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
einops>=0.6.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.62.0
